[COLLAB-HUB] Separates mapper functions fron the response classes
