+++ src/main/java/com/tool/collabhub/dto/response/ProjectResponse.java
@@ -16,24 +16,4 @@ public class ProjectResponse {
     private String description;
     private LocalDateTime createdOn;
     private String createdBy;
-
-    public static ProjectResponse from(Project project) {
-        return ProjectResponse.builder()
-                .id(project.getId())
-                .title(project.getTitle())
-                .description(project.getDescription())
-                .createdBy(project.getCreatedBy())
-                .createdOn(project.getCreatedOn())
-                .build();
-    }
-
-    public static ProjectResponse from(ProjectIndex projectIndex){
-        return ProjectResponse.builder()
-                .id(projectIndex.getId())
-                .title(projectIndex.getTitle())
-                .description(projectIndex.getDescription())
-                .createdBy(projectIndex.getCreatedBy())
-                .createdOn(projectIndex.getCreatedOn())
-                .build();
-    }
 }
